### 2025-10-28 f8c738d
Extending the supported keywords in props
