### 2025-10-28 e6372a8
Python opm now available for macOS
