+++ tests/test_6_segwell_faults.py
@@ -24,21 +24,8 @@ OPERNUM 2 30 38 47 52 2* /
 import os
 import pathlib
 import subprocess
-import warnings
 import numpy as np
 from resdata.resfile import ResdataFile
-
-has_opm = False
-try:
-    has_opm = bool(__import__("opm"))
-except ImportError:
-    warnings.warn(
-        UserWarning(
-            "The optional OPM Python package for postprocessing "
-            "the simulation files is not found; then, test_6_segwell_faults.py is only run "
-            "using resdata."
-        )
-    )
 
 testpth: pathlib.Path = pathlib.Path(__file__).parent
 mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
@@ -56,10 +43,7 @@ def test_complex(flow):
         ],
         check=True,
     )
-    readers = ["resdata"]
-    if has_opm:
-        readers += ["opm"]
-    for use in readers:
+    for use in ["opm", "resdata"]:
         sub = use.upper()
         subprocess.run(
             [
