+++ tests/test_2_coarsening.py
@@ -7,19 +7,6 @@
 import os
 import pathlib
 import subprocess
-import warnings
-
-has_opm = False
-try:
-    has_opm = bool(__import__("opm"))
-except ImportError:
-    warnings.warn(
-        UserWarning(
-            "The optional OPM Python package for postprocessing "
-            "the simulation files is not found; then, test_2_coarsening.py is only run "
-            "using resdata."
-        )
-    )
 
 testpth: pathlib.Path = pathlib.Path(__file__).parent
 mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
@@ -29,10 +16,7 @@ def test_coarsening(flow):
     """See examples/decks/HELLO_WORLD.DATA"""
     if not os.path.exists(f"{testpth}/output"):
         os.system(f"mkdir {testpth}/output")
-    readers = ["resdata"]
-    if has_opm:
-        readers += ["opm"]
-    for use in readers:
+    for use in ["opm", "resdata"]:
         sub = f"FINER{use.upper()}"
         subprocess.run(
             [
