### 2025-11-05 345d7d1
Retiring resdata
