+++ tests/test_4_submodel.py
@@ -8,7 +8,7 @@ import os
 import pathlib
 import subprocess
 import numpy as np
-from resdata.resfile import ResdataFile
+from opm.io.ecl import EclFile as OpmFile
 
 testpth: pathlib.Path = pathlib.Path(__file__).parent
 mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
@@ -18,64 +18,59 @@ def test_submodel(flow):
     """See examples/decks/MODEL0.DATA and MODEL1.DATA"""
     if not os.path.exists(f"{testpth}/output"):
         os.system(f"mkdir {testpth}/output")
-    for use in ["opm", "resdata"]:
-        for i in range(2):
-            sub = f"MODEL{i}_FINER{use.upper()}"
-            os.chdir(f"{testpth}/output")
+    for i in range(2):
+        sub = f"MODEL{i}_FINER"
+        os.chdir(f"{testpth}/output")
+        subprocess.run(
+            [
+                "pycopm",
+                "-f",
+                flow,
+                "-i",
+                f"{mainpth}/examples/decks/MODEL{i}.DATA",
+                "-o",
+                f"{testpth}/output/submodel",
+                "-g",
+                "4,4,3",
+                "-warnings",
+                "1",
+                "-m",
+                "all",
+                "-w",
+                sub,
+                "-l",
+                sub,
+            ],
+            check=True,
+        )
+        assert os.path.exists(f"{testpth}/output/submodel/{sub}.INIT")
+        assert os.path.exists(f"{testpth}/output/submodel/{sub}.EGRID")
+        os.chdir(f"{testpth}/output/submodel")
+        for j in [1, 3, 4]:
             subprocess.run(
                 [
                     "pycopm",
                     "-f",
                     flow,
                     "-i",
-                    f"{mainpth}/examples/decks/MODEL{i}.DATA",
+                    f"{sub}.DATA",
                     "-o",
-                    f"{testpth}/output/submodel",
-                    "-g",
-                    "4,4,3",
-                    "-warnings",
-                    "1",
+                    ".",
+                    "-v",
+                    "xypolygon [3,7.5] [7.5,3] [12,7.5] [7.5,12] [3,7.5]",
                     "-m",
                     "all",
                     "-w",
-                    sub,
+                    f"{sub}SUBMODEL",
                     "-l",
-                    sub,
-                    "-u",
-                    use,
+                    f"{sub}SUBMODEL",
+                    "-warnings",
+                    "1",
+                    "-p",
+                    f"{j}",
                 ],
                 check=True,
             )
-            assert os.path.exists(f"{testpth}/output/submodel/{sub}.INIT")
-            assert os.path.exists(f"{testpth}/output/submodel/{sub}.EGRID")
-            os.chdir(f"{testpth}/output/submodel")
-            for j in [1, 3, 4]:
-                subprocess.run(
-                    [
-                        "pycopm",
-                        "-f",
-                        flow,
-                        "-i",
-                        f"{sub}.DATA",
-                        "-o",
-                        ".",
-                        "-v",
-                        "xypolygon [3,7.5] [7.5,3] [12,7.5] [7.5,12] [3,7.5]",
-                        "-m",
-                        "all",
-                        "-w",
-                        f"{sub}SUBMODEL",
-                        "-l",
-                        f"{sub}SUBMODEL",
-                        "-warnings",
-                        "1",
-                        "-p",
-                        f"{j}",
-                        "-u",
-                        use,
-                    ],
-                    check=True,
-                )
[…]
