+++ src/pycopm/utils/mapping_methods.py
@@ -933,18 +933,13 @@ def handle_vicinity(dic):
     dic["wvicinity"] = []
     dic["hvicinity"] = []
     if dic["optvic"].upper() == "XYPOLYGON":
-        if dic["resdata"]:
-            coords = dic["grid"].export_position(dic["grid"].export_index())
-        else:
-            coords = []
-            for k in range(dic["zn"]):
-                for j in range(dic["yn"]):
-                    for i in range(dic["xn"]):
-                        xyz = dic["grid"].xyz_from_ijk(i, j, k, True)
-                        coords.append(
-                            [sum(xyz[0]) / 8.0, sum(xyz[1]) / 8, sum(xyz[2]) / 8]
-                        )
-            coords = np.array(coords)
+        coords = []
+        for k in range(dic["zn"]):
+            for j in range(dic["yn"]):
+                for i in range(dic["xn"]):
+                    xyz = dic["grid"].xyz_from_ijk(i, j, k, True)
+                    coords.append([sum(xyz[0]) / 8.0, sum(xyz[1]) / 8, sum(xyz[2]) / 8])
+        coords = np.array(coords)
         nxy = dic["xn"] * dic["yn"]
         nxyz = nxy * dic["zn"]
         dic["subm"] = []
@@ -1057,16 +1052,9 @@ def handle_vicinity(dic):
         found = False
         dic["subm"] = np.ones(dic["xn"] * dic["yn"] * dic["zn"])
         quans = [int(val) for val in vicinity[1].split(",")]
-        if dic["resdata"]:
-            if dic["ini"].has_kw(dic["optvic"].upper()):
-                dic["subm"][dic["actind"]] = dic["ini"].iget_kw(dic["optvic"].upper())[
-                    0
-                ]
-                found = True
-        else:
-            if dic["ini"].count(dic["optvic"].upper()):
-                dic["subm"][dic["actind"]] = dic["ini"][dic["optvic"].upper()]
-                found = True
+        if dic["ini"].count(dic["optvic"].upper()):
+            dic["subm"][dic["actind"]] = dic["ini"][dic["optvic"].upper()]
+            found = True
         if found:
             dic["subm"] = [val in quans for val in dic["subm"]]
         else:
@@ -1129,16 +1117,10 @@ def map_vicinity(dic):
             bar_animation()
             if name != "porv":
                 dic[name] = np.zeros(nc)
-                if dic["resdata"]:
-                    if name in dic["rptrst"]:
-                        dic[name][dic["actind"]] = dic["rst"].iget_kw(name.upper())[0]
-                    else:
-                        dic[name][dic["actind"]] = dic["ini"].iget_kw(name.upper())[0]
+                if name in dic["rptrst"]:
+                    dic[name][dic["actind"]] = dic["rst"][name.upper(), 0]
                 else:
-                    if name in dic["rptrst"]:
-                        dic[name][dic["actind"]] = dic["rst"][name.upper(), 0]
-                    else:
-                        dic[name][dic["actind"]] = dic["ini"][name.upper()]
+                    dic[name][dic["actind"]] = dic["ini"][name.upper()]
             dic[f"{name}_c"] = [""] * (dic["nx"] * dic["ny"] * dic["nz"])
             n = 0
             for k in range(dic["zn"]):
@@ -1278,12 +1260,8 @@ def handle_cp_grid(dic):
     """
     ir = []
     mr = []
-    if dic["resdata"]:
-        zc = dic["grid"].export_zcorn()
[…]
