+++ pyproject.toml
@@ -27,7 +27,6 @@ dependencies = [
     "ert==13.0.4",
     "h5py<=3.14.0",
     "mako<=1.3.10",
-    "resdata<=6.0.1",
     "opm<=2025.10"
 ]
 requires-python = ">=3.11"

[…]
