### 2025-11-12 e0eba38
Support for WPAVEDEP, WDFAC, and COMPLUMP
