+++ .github/workflows/ci_pycopm_ubuntu.yml
@@ -35,6 +35,10 @@ jobs:
         sudo apt-get update
         sudo apt-get install mpi-default-bin
         sudo apt-get install libopm-simulators-bin
+
+    - name: Install ert dependency
+      run: |
+        sudo apt-get install freeglut3-dev
         
     - name: Install dependecies
       run: |
