+++ src/pycopm/utils/properties_builder.py
@@ -56,7 +56,7 @@ def coarser_properties(dic):
             dic["fluxnum_c"][num] = int(min(dic["fluxnum"][inx]))
             dic["fipnum_c"][num] = int(min(dic["fipnum"][inx]))
             dic["eqlnum_c"][num] = int(min(dic["eqlnum"][inx]))
-            dic["vol_c"][num] = sum(dic["vol"][inx])
+            dic["vol_c"][num] = np.sum(dic["vol"][inx])
             if dic["field"] == "drogon":
                 dic["multnum_c"][num] = int(min(dic["multnum"][inx]))
                 dic["pvtnum_c"][num] = int(min(dic["pvtnum"][inx]))
@@ -65,12 +65,12 @@ def coarser_properties(dic):
             if dic["deck"] == 1 and (dic["letsatn"] == 1 or dic["letsatn"] == 3):
                 dic["satnum_c"][num] = int(min(dic["satnum"][inx]))
             if dic["deck"] == 1 and dic["letsatn"] == 2:
-                if sum(dic["actnum_c"][0:num]) > 1:
-                    dic["satnum_c"][num] += int(sum(dic["actnum_c"][0:num])) - 1
+                if np.sum(dic["actnum_c"][0:num]) > 1:
+                    dic["satnum_c"][num] += int(np.sum(dic["actnum_c"][0:num])) - 1
 
-            if sum(dic["actnum"][inx]) > 0:
+            if np.sum(dic["actnum"][inx]) > 0:
                 dic["poro_c"][num] = (
-                    sum(dic["poro"][inx] * dic["vol"][inx] * dic["actnum"][inx])
+                    np.sum(dic["poro"][inx] * dic["vol"][inx] * dic["actnum"][inx])
                     / dic["vol_c"][num]
                 )
                 for i, name in zip(range(3), ["permx", "permy", "permz"]):
@@ -78,11 +78,13 @@ def coarser_properties(dic):
                         dic[name + "_c"][num] = max(dic[name][inx])
                     else:
                         dic[name + "_c"][num] = (
-                            sum(dic[name][inx] * dic["vol"][inx] * dic["actnum"][inx])
+                            np.sum(
+                                dic[name][inx] * dic["vol"][inx] * dic["actnum"][inx]
+                            )
                             / dic["vol_c"][num]
                         )
                 dic["ntg_c"][num] = (
-                    sum(dic["ntg"][inx] * dic["vol"][inx] * dic["actnum"][inx])
+                    np.sum(dic["ntg"][inx] * dic["vol"][inx] * dic["actnum"][inx])
                     / dic["vol_c"][num]
                 )
                 pv_f = (
@@ -91,16 +93,16 @@ def coarser_properties(dic):
                     * dic["actnum"][inx]
                     * dic["ntg"][inx]
                 )
-                pv_c = sum(pv_f)
+                pv_c = np.sum(pv_f)
                 dic["porv_c"][num] = pv_c
-                dic["swl_c"][num] = sum(dic["swl"][inx] * pv_f) / pv_c
-                dic["sgu_c"][num] = sum(dic["sgu"][inx] * pv_f) / pv_c
-                dic["swcr_c"][num] = sum(dic["swcr"][inx] * pv_f) / pv_c
-                dic["swat_c"][num] = sum(dic["swat"][inx] * pv_f) / pv_c
-                dic["sgas_c"][num] = sum(dic["sgas"][inx] * pv_f) / pv_c
-                dic["pressure_c"][num] = sum(dic["pressure"][inx] * pv_f) / pv_c
-                dic["rs_c"][num] = sum(dic["rs"][inx] * pv_f) / pv_c
[…]
