+++ pyproject.toml
@@ -24,7 +24,7 @@ classifiers = [
 ]
 dependencies = [
     "alive_progress<=3.3.0",
-    "ert==16.0.8",
+    "ert@git+https://github.com/equinor/ert#egg=d79d7c9", # Using master since the latest available version (16.0.8) has a few issues 
     "h5py<=3.14.0",
     "mako<=1.3.10",
     "opm<=2025.10"

[…]
