### 2025-11-25 #9242 @daavid00
Hei,

Please let me know if there is something pending from my side regarding this submission.

Many thanks,

-David
