### 2025-12-03 c8119a0
Adding files to aline with GitHub community standards
