+++ SECURITY.md
@@ -1,0 +1,40 @@
+# Security Policy
+
+## Supported Versions
+
+We only provide security updates for the following versions:
+
+*   **Version 2025.04** (Current Stable)
+
+## Reporting a Vulnerability
+
+We take security vulnerabilities seriously and appreciate your efforts to responsibly disclose them. To report a vulnerability, please follow these steps:
+
+1.  **Do not open a public GitHub issue.** This could expose the vulnerability to malicious actors before a fix is available.
+2.  **Email us directly at `dmar@norceresearch.no`**.
+3.  **Provide detailed information:**
+    *   A clear description of the vulnerability.
+    *   Steps to reproduce the vulnerability.
+    *   The affected versions of the project.
+    *   Any potential impact or exploit scenarios.
+    *   Your contact information for follow-up.
+
+## Our Disclosure Policy
+
+Upon receiving a security bug report, we will:
+
+1.  **Acknowledge receipt** within 24-72 hours.
+2.  **Confirm the problem** and determine affected versions.
+3.  **Work on a fix** for all supported releases.
+4.  **Coordinate a release** with the fix.
+5.  **Publicly acknowledge your contribution** (with your permission) after the fix is released.
+
+## Responsible Disclosure
+
+We kindly request that you:
+
+*   **Do not publicly disclose the vulnerability** until we have released a fix.
+*   **Allow us a reasonable amount of time** to address the issue before public disclosure.
+*   **Do not engage in any actions** that could harm our users or systems during your research.
+
+Thank you for helping to keep our project secure.
