+++ CITATION.cff
@@ -1,0 +1,10 @@
+cff-version: "1.2.0"
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: Landa-Marbán
+  given-names: David
+  orcid: "https://orcid.org/0000-0002-3343-1005"
+title: "pycopm: An open-source tool to tailor OPM Flow 
+  geological models"
+year: 2025
+url: "https://github.com/cssr-tools/pycopm"

[…]
