### 2025-12-26 522be6b
Docs brew formulas for flow and resinsight
