+++ docs/text/installation.rst
@@ -2,7 +2,7 @@
 Installation
 ============
 
-The following steps work installing the dependencies in Ubuntu via apt-get or in macOS using brew or macports.
+The following steps work installing the dependencies in Ubuntu via apt-get or in macOS using `brew <https://brew.sh>`_ or `macports <https://www.macports.org>`_.
 While using package managers such as Anaconda, Miniforge, or Mamba might work, these are not tested.
 The supported Python versions are 3.11 to 3.13. We will update the documentation when Python3.14 is supported
 (e.g., the ert Python package is not yet available via pip install in Python 3.14).
@@ -60,6 +60,13 @@ You also need to install:
 
 * OPM Flow (https://opm-project.org, Release 2025.10 or current master branches)
 
+Binary packages
++++++++++++++++
+
+See the `downloading and installing <https://opm-project.org/?page_id=36>`_ OPM Flow online documentation for 
+instructions to install the binary packages in Ubuntu and Red Hat Enterprise Linux, and for other platforms which are
+supported either via source builds or through running a virtual machine.
+
 .. tip::
 
     See the `ci_pycopm_ubuntu.yml <https://github.com/cssr-tools/pycopm/blob/main/.github/workflows/ci_pycopm_ubuntu.yml>`_ script 
@@ -97,30 +104,40 @@ in the terminal the following lines (which in turn should build flow in the fold
 
 .. _macOS:
 
+Brew formula for macOS
+++++++++++++++++++++++
+For macOS, there are no available binary packages, so OPM Flow needs to be built from source. Recently, a formula to build flow using brew has
+been added in `https://github.com/cssr-tools/homebrew-opm <https://github.com/cssr-tools/homebrew-opm>`_. 
+Then, you can try to install flow (v2025.10) by simply typing:
+
+.. code-block:: console
+
+    brew install cssr-tools/opm/opm-simulators
+
+You can check if the installation of OPM Flow succeded by typing in the terminal **flow \-\-help**.
+
+.. tip::
+    See the actions in the `cssr-tools/homebrew-opm <https://github.com/cssr-tools/homebrew-opm/actions>`_ repository.
+
 Source build in macOS
 +++++++++++++++++++++
-For macOS, there are no available binary packages, so OPM Flow needs to be built from source, in addition to the dune libraries 
-(see the `prerequisites <https://opm-project.org/?page_id=239>`_, which can be installed using macports or brew). For example,
-with brew the prerequisites can be installed by:
+If you would like to build the latest OPM Flow from the master branch, then you can first install the prerequisites using brew:
 
 .. code-block:: console
 
     brew install boost openblas suite-sparse python@3.13 cmake 
 
 In addition, it is recommended to uprade and update your macOS to the latest available versions (the following steps have 
-worked for macOS Tahoe 26.0.1 with Apple clang version 17.0.0).
-After the prerequisites are installed and the vpyocpm Python environment is created (see :ref:`vpycopm`), 
-then building OPM Flow can be achieved with the following bash lines:
+worked for macOS Tahoe 26.2.0 with Apple clang version 17.0.0). After the prerequisites are installed, then building OPM Flow
+can be achieved with the following bash lines:
 
 .. code-block:: console
 
     CURRENT_DIRECTORY="$PWD"
 
-    deactivate
-    source vpycopm/bin/activate
-
     for module in common geometry grid istl
[…]
