+++ docs/text/conf.py
@@ -7,7 +7,7 @@
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 project = 'pycopm'
-copyright = '2024-2025, NORCE Research AS'
+copyright = '2024-2026, NORCE Research AS'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
