+++ src/pycopm/utils/generate_files.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2024-2025 NORCE Research AS
+# SPDX-FileCopyrightText: 2024-2026 NORCE Research AS
 # SPDX-License-Identifier: GPL-3.0
 # pylint: disable=R0912,R0913,R0914,R0915,C0302,R0917,R1702,R0916,R0911,R0801,E1102
 
@@ -695,7 +695,7 @@ def write_grid(dic):
     grid.append(
         "-- This deck was generated by pycopm https://github.com/cssr-tools/pycopm\n"
     )
-    grid.append("-- Copyright (C) 2024-2025 NORCE Research AS\n")
+    grid.append("-- Copyright (C) 2024-2026 NORCE Research AS\n")
     grid.append("SPECGRID\n")
     grid.append(f"{dic['nx']} {dic['ny']} {dic['nz']} /\n")
     grid.append("COORD\n")
