+++ tests/test_4_submodel.py
@@ -32,8 +32,6 @@ def test_submodel(flow):
                 f"{testpth}/output/submodel",
                 "-g",
                 "4,4,3",
-                "-warnings",
-                "1",
                 "-m",
                 "all",
                 "-w",
@@ -64,8 +62,6 @@ def test_submodel(flow):
                     f"{sub}SUBMODEL",
                     "-l",
                     f"{sub}SUBMODEL",
-                    "-warnings",
-                    "1",
                     "-p",
                     f"{j}",
                 ],
