+++ tests/test_6_segwell_faults.py
@@ -103,8 +103,6 @@ def test_complex(flow):
                 f"{testpth}/output/complex",
                 "-i",
                 f"{mainpth}/examples/decks/MODEL3.DATA",
-                "-warnings",
-                "1",
                 "-g",
                 "2,2,2",
                 "-w",
@@ -153,8 +151,6 @@ def test_complex(flow):
                 f"A4 {val}",
                 "-w",
                 f"SUBMODEL{i}",
-                "-warnings",
-                "1",
                 "-l",
                 f"S{i}",
                 "-p",
