+++ examples/configurations/drogon/input.toml
@@ -3,7 +3,7 @@ flow = "flow"
 
 # Set the model parameters
 field = "drogon"  # Geological model (norne or drogon)
-mode = "single-run" # Mode to run (single-run, files, or ert)
+mode = "ert" # Mode to run (single-run, files, or ert)
 X = [0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0] # Array of x-coarsening 
 Y = [0,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0] # Array of y-coarsening 
 Z = [0,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0,2,0] # Array of z-coarsening
