### 2026-01-15 4c6bea6
Update regarding JOSS scope and paper format changes
