+++ paper/paper.bib
@@ -33,7 +33,7 @@ abstract = {The Open Porous Media (OPM) initiative is a community effort that en
 }
 
 @software{the_pandas_development_team,
-  author       = {The pandas development team},
+  author       = {{The pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = sep,
   year         = 2025,
@@ -73,3 +73,72 @@ year = {2025}
    publisher = "European Association of Geoscientists &amp; Engineers",
    issn = "2214-4609"
   }
+
+@article{nilsen2025,
+title = {Closed-loop workflow for short-term optimization of wind-powered reservoir management},
+journal = {Cleaner Energy Systems},
+volume = {12},
+pages = {100213},
+year = {2025},
+issn = {2772-7831},
+doi = {https://doi.org/10.1016/j.cles.2025.100213},
+url = {https://www.sciencedirect.com/science/article/pii/S2772783125000445},
+author = {Mathias M. Nilsen and Rolf J. Lorentzen and Olwijn Leeuwenburgh and Andreas S. Stordal and Eduardo Barros},
+keywords = {Reservoir management, Multi-scale optimization, Wind power, Closed-loop},
+abstract = {This paper presents a closed-loop workflow for short-term optimization of reservoir management powered by offshore wind energy. Motivated by the need to reduce CO2 emissions in the Norwegian oil and gas sector, the workflow integrates wind power forecasts into optimization of daily well control while adhering to a long-term production strategy optimized for economic output. The workflow utilizes coarsened reservoir models calibrated through ensemble-based data assimilation to minimize computational costs. A realistic benchmark model, Drogon, is used to demonstrate the methodology. The workflow dynamically adjusts well rates to align power demand with wind availability, minimizing reliance on gas turbines and thereby reducing emissions. The numerical experiment demonstrates the potential of the workflow by significantly reducing short-term emissions without compromising the NPV.}
+}
+
+@misc{landamarbán2025,
+      title={A Coarsening Approach to the Troll Aquifer Model}, 
+      author={David Landa-Marbán and Tor Harald Sandve and Sarah Eileen Gasda},
+      year={2025},
+      eprint={2508.08670},
+      archivePrefix={arXiv},
+      primaryClass={physics.geo-ph},
+      url={https://arxiv.org/abs/2508.08670}, 
+}
+
+@article{Sandve2025,
+   author = "Sandve, T.H. and Boon, W. and Landa-Marbán, D. and Tveit, S. and Gasda, S.E.",
+   title = "Multi-Scale Simulation Strategies for Managing Pressure Interference in Multi-Site CO2 Storage in Large Regional Aquifers", 
+   journal= "",
+   year = "2025",
+   volume = "2025",
+   number = "1",
+   pages = "1-4",
+   doi = "https://doi.org/10.3997/2214-4609.202521134",
+   url = "https://www.earthdoc.org/content/papers/10.3997/2214-4609.202521134",
+   publisher = "European Association of Geoscientists &amp; Engineers",
+   issn = "2214-4609",
+   type = "",
+  }
+
+@misc{liu2025,
+      title={Trends in porous media laboratory imaging and open science practices}, 
+      author={Na Liu and Jakub Wiktor Both and Geir Ersland and Jan Martin Nordbotten and Martin Fernø},
+      year={2025},
+      eprint={2510.05190},
+      archivePrefix={arXiv},
+      primaryClass={physics.flu-dyn},
+      url={https://arxiv.org/abs/2510.05190}, 
+}
+
+
+@article{Wilkinson2016,
[…]
