### 2026-01-24 #9242 @daavid00
Hi @kthyng, thanks for checking the repository. I have updated the paper in https://github.com/cssr-tools/pycopm/pull/72 to add the `State of the field` section, and also updated the `Software design` section since it seems there has been changes in the [JOSS scope definition](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance) after I updated the paper last week. I look forward to your evaluation, and thanks again for your time.
