### 2026-01-28 122d703
Updates to the JOSS paper
