### 2026-01-28 #9242 @daavid00
Hi @arfon, many thanks for sharing the new requirements, and congratulations to all the JOSS team for coming with this improved new format 🙂. I have updated the paper in https://github.com/cssr-tools/pycopm/pull/74, aiming to comply with these new requirements and format.
