### 2026-02-13 cad9232
Adding arXiv preprint of the repository paper
