### 2026-02-18 f11068a
Reducing the output size
