### 2026-02-19 5b6ee4f
Bug fix for coarsening and -t 1 or -t 2
