### 2026-02-20 #9242 @kthyng
@daavid00 Congrats! Your submission has passed all of my checks! This is the first one to clearly do so for me (since the new scope requirements) so this is exciting. I will add it to the waitlist and see about getting you an editor
