### 2026-02-23 #9242 @andreww
Happy to edit this @kthyng.

@daavid00 - does the list of potential reviewers from October (above) still look good to you? If you have any further suggestions  please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

The start of this week looks quite hectic here: I'll probably not get as far as kicking off the search for reviewers until Wednesday.
