### 2026-02-25 #9242 @andreww
:wave: @lachlangrose, @frank1010111, @santisoler, @nasserma, @atgeirr, @bska, @akva2 would any of you be willing to review this submission ("pycopm: An open-source tool to tailor OPM Flow geological models") for the Journal of Open Source Software?

We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
