### 2026-02-26 #9242 @bska
I too believe that I have a conflict of interest here since I'm a technical maintainer of several of the repositories to which the author contributes.
