### 2026-02-27 #10124 @andreww
:wave: @daavid00 @frank1010111 @nasserma this is the review thread for the paper. All of our communications will happen here from now on.

I see you've both reviewed for JOSS before so my guidance below is probably familiar. However, it's worth pointing out that the scope requirements changed at the start of this year - see the [announcement](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future) and [updated scope](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance). Overall, I think there is a slight increase in the emphasis on the "why do it like this" in the paper and modest changes to the expectations for the software. Do let me know if you have questions or concerns around these changes.

```@editorialbot generate my checklist```
