### 2026-03-14 582696e
Updating docs to install OPM Flow in macOS
