### 2026-03-23 #10124 @andreww
👋 @frank1010111, @nasserma - I'm just wondering how the review of pycopm is going and if you need anything from me?

@nasserma: it would be helpful if you could ask the editorialbot to generate your checklist by creating a comment `@editorialbot generate my checklist` below.

Do let me know if you have any questions or concerns.
