### 2026-03-23 #10124 @frank1010111
I'm currently dealing with the consequences of my past actions, specifically using Arch as my Linux of choice. It makes installing OPM flow difficult, and building from source has not yet worked. I will continue trying to figure that out, but I'm also open to using a docker container if you have an image that you have tested.
