### 2026-03-26 #10124 @andreww
> I'm currently dealing with the consequences of my past actions, specifically using Arch as my Linux of choice. It makes installing OPM flow difficult, and building from source has not yet worked. I will continue trying to figure that out, but I'm also open to using a docker container if you have an image that you have tested.

Yes - that looks non-trivial. I wonder if @daavid00 has any suggestions. I imagine reviewing pycomp shouldn't need an optimised version of OPM flow, for example.
