### 2026-03-27 #10124 @daavid00
> > I'm currently dealing with the consequences of my past actions, specifically using Arch as my Linux of choice. It makes installing OPM flow difficult, and building from source has not yet worked. I will continue trying to figure that out, but I'm also open to using a docker container if you have an image that you have tested.
>
> Yes - that looks non-trivial. I wonder if [@daavid00](https://github.com/daavid00) has any suggestions. I imagine reviewing pycomp shouldn't need an optimised version of OPM flow, for example.

Hei, [this repository](https://github.com/daavid00/OPM-Flow_macOS) builds OPM Flow from source using the master branches with support for mpirun in the latest available macOS in GitHub Actions, showing the status and details of the build in the [Actions](https://github.com/daavid00/OPM-Flow_macOS/actions). Then maybe this can be of help for the needed requirements/cmake flags to build it in Arch. Unfortunately I do not have a machine with Arch to test this, and also I have not used docker; however, there might be developers/users that have manage this in the OPM community and can help, then please [raise an issue](https://github.com/OPM/opm-simulators/issues/new) in opm-simulators with the error you get in Arch. I will also look at the docker container option, and let you know if I find a solution.

As @andreww commented, reviewing pycopm does not required an optimised version of OPM flow, and also does not required a parallel version (mpi support) as well.
