### 2026-04-03 #10124 @nasserma
@daavid00 OPM Flow has binary packages for for Ubuntu 24.04, but I am using 25.10 and would need to compile. I am fine with this, but can you confirm which modules are needed (opm-common, opm-grid, opm-simulators, opm-upscaling)?
