### 2026-04-03 #10124 @nasserma
@daavid00 @andreww I have installed the package, but cannot run the unit tests or examples until I compile OPM Flow, which I will do once the installation documentation is updated to explain which modules are needed (or that all of them are).

I do have one initial high-level comment/question, this project seems to only have one contributor, so it would not meet the _collaborative effort_ criteria listed in the review template. Is this a hard constraint for JOSS acceptance?
