### 2026-04-03 #10124 @daavid00
Hei, only opm-common, opm-grid, and opm-simulators are needed. In the [installation section](https://cssr-tools.github.io/pycopm/installation.html#source-build-in-linux-windows) of the docs the building from source in Linux is described. Please let me know if you have any issues.
