### 2026-04-08 #10124 @andreww
> I do have one initial high-level comment/question, this project seems to only have one contributor, so it would not meet the _collaborative effort_ criteria listed in the review template. Is this a hard constraint for JOSS acceptance?

Hi @nasserma - apologies, I've been away over the long weekend here. In terms of collaborative effort, it is possible for single contributor packages to meet this criteria. It is okay for a single author project to show "other evidence of community engagement, either in the repository or evidenced in the paper". Things to look for include requests for and discussions of new features (or bug reports or other improvements). This may be clear from the repository / issue tracker / commit history but it's also okay for this to be evidenced in the paper itself. See https://joss.readthedocs.io/en/latest/review_criteria.html#collaborative-effort and it's okay to ask for more evidence of this as part of the review process.
