### 2026-04-09 #10124 @daavid00
> I'm currently dealing with the consequences of my past actions, specifically using Arch as my Linux of choice. It makes installing OPM flow difficult, and building from source has not yet worked. I will continue trying to figure that out, but I'm also open to using a docker container if you have an image that you have tested.

Here is the link to the online documentation for the docker:
https://opm-project.org/?page_id=853

and the docker:
https://hub.docker.com/r/openporousmedia/opmreleases

I have not tested this personally, but I have asked to a college who has and he said it works.
