### 2026-04-15 #10124 @andreww
Hi @frank1010111, @nasserma and @daavid00 - how are things going with the review / getting the code to run for you? I'm mostly going to be away for the rest of this week and all of next week so won't be proactively checking this issue but if you tag me here I'll try and pick anything up. Do tag any related issues you add to the pycopm (or other OMP repos) with this issue to help keep track of things (including a link to `https://github.com/openjournals/joss-reviews/issues/10124` does this).
