### 2026-04-15 #10124 @daavid00
Hi @andreww, thanks for following this review. @frank1010111 and @nasserma, if you have issues installing OPM Flow, please raise an issue at https://github.com/OPM/opm-simulators/issues with details on the logs, and I will follow this with other OPM Flow developers.
