### 2026-04-28 #10124 @andreww
@frank1010111 and @nasserma - could you give a brief update on the state of your reviews of pycomp and if you've run into any problems please? I see @nasserma has many of the items ticked off the list, but @frank1010111 doesn't. Is the issue still installation?
