### 2026-05-11 437849a
JOSS review: paper suggestions by frank1010111
