### 2026-05-20 9907765
:whale: Dockerfile for reproducible test environment
