### 2026-05-20 3a1e5ff
:whale: add freeglut3-dev dependency to dockerfile for ert
