+++ pyproject.toml
@@ -20,15 +20,15 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "Operating System :: OS Independent"
 ]
 dependencies = [
     "alive_progress<=3.3.0",
-    "ert<=17.0.0",
+    "ert==21.1.5",
     "h5py<=3.14.0",
-    "iterative_ensemble_smoother<=0.6.0",
     "mako<=1.3.10",
-    "opm<=2025.10"
+    "opm<=2026.4"
 ]
 requires-python = ">=3.11"
 
