### 2026-05-20 3d23b4a
Updating the test to use tmp_path and monkeypatch
