+++ tests/test_3_refinement.py
@@ -1,33 +1,26 @@
 # SPDX-FileCopyrightText: 2025-2026 NORCE Research AS
 # SPDX-License-Identifier: GPL-3.0
-# pylint: disable=R0801
 
 """Test the generic refinement functionality"""
 
-import os
-import pathlib
+from pathlib import Path
 import subprocess
 import numpy as np
 from opm.io.ecl import EclFile as OpmFile
 
-mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
-testpth: pathlib.Path = pathlib.Path(__file__).parent
 
-
-def test_refinement(flow):
+def test_3_refinement(flow, tmp_path, monkeypatch):
     """See examples/decks/MODEL2.DATA"""
-    if not os.path.exists(f"{testpth}/output"):
-        os.system(f"mkdir {testpth}/output")
+    monkeypatch.chdir(tmp_path)
+    repo_root = Path(__file__).parents[1]
     sub = "FINER"
     subprocess.run(
         [
             "pycopm",
+            "-i",
+            f"{repo_root}/examples/decks/MODEL2.DATA",
             "-f",
             flow,
-            "-i",
-            f"{mainpth}/examples/decks/MODEL2.DATA",
-            "-o",
-            f"{testpth}/output/finer",
             "-g",
             "2,4,8",
             "-l",
@@ -39,9 +32,8 @@ def test_refinement(flow):
         ],
         check=True,
     )
-    assert os.path.exists(f"{testpth}/output/finer/{sub}.INIT")
-    assert os.path.exists(f"{testpth}/output/finer/{sub}.EGRID")
-    os.chdir(f"{testpth}/output/finer")
+    assert (tmp_path / f"{sub}.INIT").is_file()
+    assert (tmp_path / f"{sub}.EGRID").is_file()
     porv = np.array(OpmFile(f"{sub}.INIT")["PORV"])
     assert abs(np.sum(porv) - 6506.25) < 1e-4
     assert np.sum(porv > 0) == 2430
