+++ tests/test_5_transform.py
@@ -1,22 +1,17 @@
 # SPDX-FileCopyrightText: 2025-2026 NORCE Research AS
 # SPDX-License-Identifier: GPL-3.0
-# pylint: disable=R0801
 
 """Test the generic affine transformation functionality"""
 
-import os
-import pathlib
+from pathlib import Path
 import subprocess
 from opm.io.ecl import EGrid as OpmGrid
 
-testpth: pathlib.Path = pathlib.Path(__file__).parent
-mainpth: pathlib.Path = pathlib.Path(__file__).parents[1]
 
-
-def test_transform(flow):
+def test_5_transform(flow, tmp_path, monkeypatch):
     """See examples/decks/MODEL0.DATA"""
-    if not os.path.exists(f"{testpth}/output"):
-        os.system(f"mkdir {testpth}/output")
+    repo_root = Path(__file__).parents[1]
+    monkeypatch.chdir(tmp_path)
     flags = ["translate", "scale", "rotatexy", "rotatexz", "rotateyz"]
     values = ["[5,-3,2]", "[2,.5,4]", "45", "-10", "15"]
     checks = [3, 4, 1, 3.5899, 4.848]
@@ -25,12 +20,10 @@ def test_transform(flow):
         subprocess.run(
             [
                 "pycopm",
+                "-i",
+                f"{repo_root}/examples/decks/MODEL0.DATA",
                 "-f",
                 flow,
-                "-i",
-                f"{mainpth}/examples/decks/MODEL0.DATA",
-                "-o",
-                f"{testpth}/output/transform",
                 "-d",
                 f"{flag} {val}",
                 "-m",
@@ -42,9 +35,9 @@ def test_transform(flow):
             ],
             check=True,
         )
-        assert os.path.exists(f"{testpth}/output/transform/{sub}.INIT")
-        assert os.path.exists(f"{testpth}/output/transform/{sub}.EGRID")
-        grid = OpmGrid(f"{testpth}/output/transform/{sub}.EGRID")
+        assert (tmp_path / f"{sub}.INIT").is_file()
+        assert (tmp_path / f"{sub}.EGRID").is_file()
+        grid = OpmGrid(f"{sub}.EGRID")
         dim = grid.dimension
         zcoord = grid.xyz_from_ijk(dim[0] - 1, dim[1] - 1, dim[2] - 1)[-1][-1]
         assert abs(zcoord - check) < 1e-2
