### 2026-06-06 #10124 @daavid00
Hi @frank1010111, I have successfully run a GitHub action with a modified version of your Dockerfile, please see https://github.com/cssr-tools/pycopm/pull/86#issuecomment-4637937647
