### 2026-06-06 f6842ad
:whale: follow David's suggestions for dockerfile
