### 2026-06-06 0ca6175
:whale: add libxkbcommon for pyqt
