### 2026-06-17 b5dccd6
Bug fixes for dual porosity
