### 2026-07-24 35dd62c
Updating GitHub actions (Ubuntu 26.04) and  --target-version py314 for black
