+++ .github/workflows/ci_pycopm_macos.yml
@@ -9,13 +9,11 @@ jobs:
   run-pycopm-macos:
     permissions:
       contents: read
-      pull-requests: write
     timeout-minutes: 90
-    runs-on: macos-latest
+    runs-on: macos-26.04
 
     steps:
-    - name: Checkout repository
-      uses: actions/checkout@v4
+    - uses: actions/checkout@v5
     
     - name: Install OPM Flow
       run: |
@@ -23,13 +21,13 @@ jobs:
         brew trust cssr-tools/opm
         brew install cssr-tools/opm/opm-simulators -y
 
-    - name: Install python requirements
+    - name: Install pycopm and python requirements
       run: |
         python3 -m venv vpycopm
         . vpycopm/bin/activate
         echo "$PWD/vpycopm/bin" >> $GITHUB_PATH
         pip install --upgrade pip setuptools wheel
-        pip install -e .
+        pip install .
         pip install -r dev-requirements.txt
 
     - name: Run the hello world example in pycopm
