### 2026-07-24 6b70e0f
macos-26 instead of macos-26.04
