+++ tests/test_7_transmissibilities.py
@@ -1,78 +1,194 @@
 # SPDX-FileCopyrightText: 2025-2026 NORCE Research AS
 # SPDX-License-Identifier: GPL-3.0
-# pylint: disable=R0914
+# pylint: disable=R0914,R0801
 
-"""Test the coarsening by transmissibilities"""
+"""Test the coarsening by transmissibilities."""
 
 from pathlib import Path
-import subprocess
+
 import numpy as np
+import pytest
 from opm.io.ecl import EclFile as OpmFile
+from opm.io.ecl import EGrid as OpmGrid
+
+from pycopm.core.pycopm import main
+from .utils import assert_grid_and_init
+
+RTOL = 1e-5
+ATOL = 1e-8
+
+REGRESSION_CASES = {
+    "COARSER1": {
+        "dimensions": (3, 3, 5),
+        "active": 43,
+        "checks": [
+            ("PORV", np.sum, 96214656.0),
+            ("PORV", np.min, 0.0),
+            ("PORV", np.max, 5929101.0),
+            ("DZ", np.sum, 2747.1123046875),
+            ("DZ", np.min, 10.770000457763672),
+            ("DZ", np.max, 145.52499389648438),
+            ("PERMX", np.sum, 96.75543212890625),
+            ("PERMX", np.max, 96.75543212890625),
+            ("TRANX", np.sum, 827.3911743164062),
+            ("TRANX", np.max, 96.01197052001953),
+            ("TRANY", np.sum, 968.17822265625),
+            ("TRANY", np.max, 93.61161804199219),
+            ("TRANZ", np.sum, 5008.13232421875),
+            ("TRANZ", np.max, 933.578125),
+            ("TRANNNC", np.sum, 13739.814453125),
+            ("TRANNNC", np.min, 0.4228149354457855),
+            ("TRANNNC", np.max, 25.62749671936035),
+        ],
+        "exact": [
+            ("SATNUM", np.sum, 43),
+            ("FIPNUM", np.sum, 43),
+        ],
+        "data": {
+            "length": 81,
+            "welspecs": (2, 2),
+            "compdat": ("INJ00", 2, 2, 1),
+        },
+    },
+    "COARSER2": {
+        "dimensions": (3, 3, 5),
+        "active": 43,
+        "checks": [
+            ("PORV", np.sum, 96214656.0),
+            ("PORV", np.min, 0.0),
+            ("PORV", np.max, 5929101.0),
+            ("DZ", np.sum, 2747.1123046875),
+            ("DZ", np.min, 10.770000457763672),
+            ("DZ", np.max, 145.52499389648438),
+            ("PERMX", np.sum, 96.75543212890625),
+            ("PERMX", np.max, 96.75543212890625),
+            ("TRANX", np.sum, 827.3911743164062),
+            ("TRANX", np.max, 96.01197052001953),
+            ("TRANY", np.sum, 968.17822265625),
+            ("TRANY", np.max, 93.61161804199219),
+            ("TRANZ", np.sum, 8564.0341796875),
+            ("TRANZ", np.max, 4563.7099609375),
+            ("TRANNNC", np.sum, 13739.814453125),
+            ("TRANNNC", np.min, 0.4228149354457855),
+            ("TRANNNC", np.max, 25.62749671936035),
+        ],
+        "exact": [
+            ("SATNUM", np.sum, 43),
+            ("FIPNUM", np.sum, 43),
+        ],
+        "data": {
+            "length": 81,
+            "welspecs": (2, 2),
+            "compdat": ("INJ00", 2, 2, 1),
+        },
+    },
+    "COARSER_MODEL5": {
+        "dimensions": (4, 1, 1),
+        "active": 4,
[…]
