+++ tests/test_8_dual.py
@@ -28,6 +28,7 @@ ATOL = 1e-8
                 "PORV": np.array([6.25e8, 1.25e9, 6.25e8, 6.25e8, 6.25e8, 6.25e8, 1.25e9, 6.25e8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2.50e8, 2.50e8, 2.50e8, 2.50e8, 0, 0]),
                 "TRANX": np.array([3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 0, 3.410807e-2, 3.410807e-2, 3.410807e-2, 0]),
                 "TRANY": np.zeros(12),
+                "TRANZ": np.zeros(12),
                 "TRANNNC": np.array([0.0682094, 0.0682094]),
                 "NNC1": np.array([2, 7]),
                 "NNC2": np.array([19, 22]),
@@ -41,9 +42,38 @@ ATOL = 1e-8
                 "PORV": np.array([6.25e8, 1.25e9, 6.25e8, 6.25e8, 6.25e8, 6.25e8, 1.25e9, 6.25e8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2.50e8, 2.50e8, 2.50e8, 2.50e8, 0, 0]),
                 "TRANX": np.array([3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 3.456000e2, 0, 3.410807e-2, 3.410807e-2, 3.410807e-2, 0]),
                 "TRANY": np.zeros(12),
+                "TRANZ": np.zeros(12),
                 "TRANNNC": np.array([0.0682094, 2.5578527, 2.5578527, 2.5578527, 2.5578527, 0.0682094]),
                 "NNC1": np.array([2, 3, 4, 5, 6, 7]),
                 "NNC2": np.array([19, 19, 20, 21, 22, 22]),
+            },
+        ),
+        (
+            "MODEL6.DATA",
+            "1:2,3:4",
+            "poro <= 0.1",
+            {
+                "PORV": np.array([625000000.0, 625000000.0, 312500000.0, 312500000.0, 312500000.0, 312500000.0, 625000000.0, 625000000.0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 125000000.0, 125000000.0, 125000000.0, 125000000.0, 0, 0, 0, 625000000.0, 312500000.0, 312500000.0, 312500000.0, 312500000.0, 625000000.0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 125000000.0, 125000000.0, 125000000.0, 125000000.0, 0, 0]),
+                "TRANX": np.array([172.8000030517578, 172.8000030517578, 172.8000030517578, 172.8000030517578, 172.8000030517578, 172.8000030517578, 172.8000030517578, 0, 0.017054030671715736, 0.017054030671715736, 0.017054030671715736, 0, 172.8000030517578, 172.8000030517578, 172.8000030517578, 172.8000030517578, 172.8000030517578, 0, 0.017054030671715736, 0.017054030671715736, 0.017054030671715736, 0]),
+                "TRANY": np.zeros(22),
+                "TRANZ": np.array([0, 720.0, 0, 0, 0, 0, 720.0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),
[…]
