+++ src/pycopm/utils/vicinity.py
@@ -1,0 +1,1129 @@
+# SPDX-FileCopyrightText: 2024-2026 NORCE Research AS
+# SPDX-License-Identifier: GPL-3.0
+# pylint: disable=R0902,R0912,R0913,R0914,R0915,C0302,R0917,R1702,R0916,R0911,E1102
+
+"""Extract submodels and map pore volume from outside their boundaries."""
+
+import csv
+import sys
+from contextlib import nullcontext
+from dataclasses import dataclass, field
+from pathlib import Path
+
+import numpy as np
+from alive_progress import alive_bar
+from numpy.typing import NDArray
+from shapely import Polygon, contains_xy, prepare
+
+from pycopm.config.config import ConfigViaDeck
+from pycopm.utils.files_writer import write_grid, write_property_inc
+
+
+@dataclass(slots=True)
+class VicinityMaps:
+    """Store a vicinity selection, bounds, and pore-volume mappings.
+
+    Global and per-layer bounds are one-based and inclusive.
+    """
+
+    #: Selected well name, or an empty string for polygon and region
+    #: selections.
+    selector: str
+
+    #: Well-vicinity shape: ``box``, ``diamond``, or ``diamondxy``. ``None`` is
+    #: used for polygon and region selections.
+    shape: str | None
+
+    #: Boolean mask identifying selected original-grid cells, flattened in
+    #: ``(z, y, x)`` order.
+    cell_mask: NDArray
+
+    #: Minimum selected i index across all layers.
+    min_i: int
+
+    #: Maximum selected i index across all layers.
+    max_i: int
+
+    #: Minimum selected j index across all layers.
+    min_j: int
+
+    #: Maximum selected j index across all layers.
+    max_j: int
+
+    #: Minimum selected k index.
+    min_k: int
+
+    #: Maximum selected k index.
+    max_k: int
+
+    #: Minimum selected i index in each original layer.
+    layer_min_i: NDArray
+
+    #: Maximum selected i index in each original layer.
+    layer_max_i: NDArray
+
+    #: Minimum selected j index in each original layer.
+    layer_min_j: NDArray
+
+    #: Maximum selected j index in each original layer.
+    layer_max_j: NDArray
+
+    #: Total pore volume of selected active cells in each original layer.
+    layer_selected_porv: NDArray
+
+    #: Total pore volume outside the selection in each original layer.
+    layer_external_porv: NDArray
+
+    #: Zero-based completion coordinates in ``[i, j, k]`` order for the
+    #: selected well.
+    well_cells: list[list[int]]
+
+    #: Zero-based output-grid indices of selected well cells. These cells are
+    #: excluded from boundary source remapping.
+    well_indices: list[int] = field(default_factory=list)
+
+    #: Number of selected active cells in each output layer.
+    active_counts: NDArray = field(default_factory=lambda: np.array([], dtype=int))
+
+    #: Original-grid source index assigned to each output boundary cell during
+    #: pore-volume correction. Zero denotes no assigned source.
+    source_indices: NDArray = field(default_factory=lambda: np.array([], dtype=int))
+
+
+@dataclass(slots=True)
+class _BoundaryMapping:
+    """Store pore-volume mapping results for one submodel boundary."""
+
+    #: Unassigned pore volume collected along the boundary.
+    pore_volume: float
+
+    #: Number of active cells receiving pore volume from the boundary.
+    active_count: int
+
+    #: Distance from the geometric boundary to each receiving cell.
+    offsets: NDArray
+
+
+def _submodel_index(dck: ConfigViaDeck, column: int, row: int, layer: int) -> int:
+    return column + row * dck.output_nx + layer * dck.output_nx * dck.output_ny
+
+
+def _original_index(dck: ConfigViaDeck, column: int, row: int, layer: int) -> int:
[…]
