### 2026-08-13 ffbe4ba
Correcting example with missing endpoint for the xypoygon
