### 2026-08-17 c1d5c6e
Improving the cli parsing
