### 2026-08-18 #10124 @andreww
>  To achieved this, I have been assisted by Copilot, then I have updated the [AI Disclosure ](https://github.com/cssr-tools/pycopm/blob/main/paper/paper.md#ai-usage-disclosure) in the paper. [@andreww](https://github.com/andreww), please let me know (when you are back) if this is sufficient or if I need to add more details.

Yes @daavid00 - I think the updated AI use disclosure covers everything needed in enough detail. I guess one thing you may want to ask yourself (but which doesn't need an answer here, or in the manuscript) is if you are still comfortable maintaining pycomp after the modernisation.
