### 2026-08-18 #10124 @andreww
I think the only thing this needs now is for @frank1010111 to take a look at the updated API documentation and if that all looks okay to tick that last remaining box on the review checklist.
