### 2026-08-18 #10124 @daavid00
> I guess one thing you may want to ask yourself (but which doesn't need an answer here, or in the manuscript) is if you are still comfortable maintaining pycomp after the modernisation.

Yes, the modernization has substantially improve readability and performance, so now it should be easier to extend functionality, and hopefully for others to start to contribute to the project :).
