### 2026-08-24 #10124 @daavid00
Hi @andreww, I have fixed the flagged DOIs by editorialbot (the A Coarsening Approach to the Troll Aquifer Model is still in arXiv).

> * Make a tagged release of your software, and list the version tag of the archived version here. You may want to make a new release capturing all changes made during review first (if you've not already)

https://github.com/cssr-tools/pycopm/releases/tag/v2026.04

> * Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

This has been done in figshare.

> * Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

I have checked the title in figshare and the paper.md match.

> * Please list the DOI of the archived version here.

https://doi.org/10.6084/m9.figshare.33321138

Please let me know if something needs to be corrected.
