### 2026-08-24 #10124 @daavid00
> [@daavid00](https://github.com/daavid00) - could you fix the licence on the Figshare archive? It's currently CC BY 4.0 but it should match the repository (GPL3). It would also be good if you could embed the version (v2026.04) in the metadata (but I'm not sure how easy that is with figshare). I think this will trigger a new version of the figshare DOI.

Ups, sorry for that. I have updated the license and added (v2026.04) at the end of the title. This has not generated a new DOI, only a new version in figshare (version 2). Please let me know if I should make further corrections.
