### 2026-08-25 #10124 @andreww
Hi @daavid00 - just some minor comments on the paper itself from me. Looks good apart from these suggestions:

- Could you include a reference to Figure 1 in the text somewhere? One option would be at the end of the sentence that ends "or changing the shape and position of the model." on line 12.
- I think "standard-industry" should be "industry-standard" on line 26.
- Should "PVT" be defined on line 29? I'm thinking pressure-volume-temperature equations of state but I could be wrong.
- "OPM Flow manual" is repeated (inside and outside the link) on lines 29 and 30.
- Are references available for Petrel, Aspen RMS, opm-upscaling or ResInsight (lines 54-75). If so, is it worth citing them?
- Should "see the theory in..." be "see the section on theory in..." on line 84?
- I think "pandas.Series.groupby" (line 89) and "contains_xy" (line 90) should both be typeset in the same way as "pycomp" is typeset. I don't think I spotted any other function names.
- The reference for "a coarsening approach to the troll aquifer model" could use the DOI [10.48550/arXiv.2508.08670](https://doi.org/10.48550/arXiv.2508.08670) which is generated by Arxiv.

If you could update the paper for those (or say why you would rather stick with what is there) then I'll be able to make the recommendation.
