### 2026-08-25 d3a1867
JOSS paper corrections
