+++ paper/paper.bib
@@ -32,23 +32,6 @@ abstract = {The Open Porous Media (OPM) initiative is a community effort that en
   doi     = {10.1038/s41586-020-2649-2}
 }
 
-@software{the_pandas_development_team,
-  author       = {{The pandas development team}},
-  title        = {pandas-dev/pandas: Pandas},
-  month        = sep,
-  year         = 2025,
-  publisher    = {Zenodo},
-  version      = {v2.3.3},
-  doi          = {10.5281/zenodo.17229934},
-  url          = {https://doi.org/10.5281/zenodo.17229934},
-  swhid        = {swh:1:dir:5bc73e541e362cf8e7068ebb434b9d7bc1194c14
-                   ;origin=https://doi.org/10.5281/zenodo.3509134;vis
-                   it=swh:1:snp:8c6e563b81f60a3d62fec309b38f470a0eacf
-                   596;anchor=swh:1:rel:b0f463d37766850a89633bd05643d
-                   45342b9241d;path=pandas-dev-pandas-1efe649
-                  }
-}
-
 @software{Gillies_Shapely_2025,
 author = {Gillies, Sean and van der Wel, Casper and Van den Bossche, Joris and Taves, Mike W. and Arnott, Joshua and Ward, Brendan C. and {others}},
 doi = {10.5281/zenodo.5597138},
@@ -95,7 +78,8 @@ abstract = {This paper presents a closed-loop workflow for short-term optimizati
       eprint={2508.08670},
       archivePrefix={arXiv},
       primaryClass={physics.geo-ph},
-      url={https://arxiv.org/abs/2508.08670}, 
+      url={https://arxiv.org/abs/2508.08670},
+      doi={10.48550/arXiv.2508.08670}
 }
 
 @article{Sandve2025,
