### 2026-08-25 #10124 @daavid00
Thanks @andreww for spotting the issues, I have addressed them in https://github.com/cssr-tools/pycopm/pull/109, with except of:
> * Are references available for Petrel, Aspen RMS, opm-upscaling or ResInsight (lines 54-75). If so, is it worth citing them?

I think for this case it is ok to put the link where it is mentioned in the text as how it is now, since I could not find official ways to cite them, and the mentions are in general and not to specific versions.

> * I think "pandas.Series.groupby" (line 89) and "contains_xy" (line 90) should both be typeset in the same way as "pycomp" is typeset. I don't think I spotted any other function names.

During the code modernization pandas was removed, then removing this also from the paper.

Please let me know if this needs more corrections, or if it is ok to merge this.
