### 2026-09-03 #10124 @kthyng
@daavid00  paper comments
- Page 1, Figure 1 caption: There is an extra closing brace at the end: “(here are details to reproduce this). }”
- Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, Landa-Marbán et al. (2025) reference: “the troll aquifer model” should probably be “the Troll aquifer model.”
