### 2026-09-04 #10124 @daavid00
Hi @kthyng, many thanks for following this contribution and for your comments, which was been merged in https://github.com/cssr-tools/pycopm/pull/111.

I also updated the reference from the arXiv preprint to the actual publication for "A coarsening approach to the Troll aquifer model", which was made available a few days ago. This is published via NVA (Norwegian Research Information Repository), then I have used the BibteX provided there https://nva.sikt.no/my-page/messages/my-messages/01a0333c3a20-aee622d2-4366-4d34-ac60-573875995685.

Please let me know if I should make further corrections.
